+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 dev:
   - Add provider for epoch summaries
   - Add withdrawals to epoch summaries table
+  - Add withdrawals to validator day summaries table
 
 0.7.4:
   - Ensure that only canonical withdrawals are included in summaries